< General > Defined class for the fieldsets. Adjusted the padding of the field sets. Removed awkward look of the buttons. Removed useless/redundant css code. Removed the single space that the PhpStorm was inserting all the time. The problem is the space before the comment at the beginning of the file. "Accordian" -> "Accordion".
< Search >
Removes error/alert messages if the accordion selection is changed.
Fixed search controller tests due to modification in the argument format of the UWI fields search.

< Export >
Created a separated .scss file for the export module.

< Visualization >
Changed button to link. Not sure if a button can be used.
Created visualization-view to centralize code.
Changed the colors of the visualization divs.
Inserted accordion in the control panel of the visualization.
Added two options in the control panel, bar and pie charts with its corresponding possible attributes to plot.
Applied animation (zooming) to open and close the visualization centre.